ANABEL-60: Made homepage. Refactored general styles.
