+++ src/main/assets/css/icomoon.css
@@ -24,6 +24,12 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
+.icon-bolt:before {
+  content: "\e901";
+}
+.icon-times:before {
+  content: "\e900";
+}
 .icon-google-plus:before {
   content: "\ea8b";
 }
