+++ src/main/assets/js/tag.js
@@ -1,0 +1,6 @@
+(function ($, window) {
+    $('.tag:not(a)').on('click', function (event) {
+        let link = $(this).data('href');
+        $(location).attr('href', link);
+    })
+})(jQuery, window);
