+++ src/main/assets/scss/_buttons.scss
@@ -1,0 +1,5 @@
+@each $color, $value in $theme-colors {
+  .btn-transparent-#{$color} {
+    @include button-transparent-variant(rgba($value, 0.1));
+  }
+}
