+++ src/main/assets/scss/_common.scss
@@ -6,7 +6,8 @@
 
     li {
       a {
-        background-color: $blue-100;
+        color: $black;
+        background-color: $blue-600;
         margin-right: .75rem;
         font-size: 1.5rem;
 
@@ -14,7 +15,7 @@
         @include border-radius(50%);
 
         &:hover {
-          background-color: $blue-200;
+          background-color: darken($blue-600, 5%);
           cursor: pointer;
         }
       }
