+++ src/main/assets/scss/_images.scss
@@ -1,0 +1,60 @@
+%tmpl-thumb-container {
+  display: block;
+  position: relative;
+  z-index: 0;
+
+  &:before {
+    content: "";
+    display: block;
+    padding-top: 100%;
+  }
+
+  .img-wrap {
+    align-items: center;
+    bottom: 0;
+    display: flex;
+    justify-content: center;
+    height: 100%;
+    left: 0;
+    max-height: 100%;
+    max-width: 100%;
+    position: absolute;
+    right: 0;
+    top: 0;
+    width: 100%;
+  }
+
+  img {
+    flex-shrink: 0;
+    height: auto;
+    max-height: inherit;
+    max-width: inherit;
+    object-fit: contain;
+    object-position: center;
+    width: auto;
+  }
+}
+
+%tmpl-thumb-container-cover {
+  overflow: hidden;
+
+  .img-wrap {
+    align-items: center; // flex-start;
+  }
+
+  img {
+    max-height: none;
+    min-height: 100%;
+    min-width: 100%;
+    object-fit: cover;
+    object-position: center; // top center;
+  }
+}
+
+.thumb-container {
+  @extend %tmpl-thumb-container;
+
+  &.thumb-container-cover {
+    @extend %tmpl-thumb-container-cover;
+  }
+}
