+++ src/main/assets/scss/_mixins.scss
@@ -1,1 +1,2 @@
 @import "mixins/_fonts";
+@import "mixins/_buttons";
