+++ src/main/assets/scss/style.scss
@@ -1,6 +1,7 @@
 @import "../../../../node_modules/bootstrap/scss/functions";
 @import "../../../../node_modules/bootstrap/scss/variables";
 @import "../../../../node_modules/bootstrap/scss/mixins";
+@import "functions";
 @import "variables";
 @import "mixins";
 //@import "../../../../node_modules/bootstrap/scss/root";
@@ -19,7 +20,7 @@
 @import "../../../../node_modules/bootstrap/scss/custom-forms";
 @import "../../../../node_modules/bootstrap/scss/nav";
 @import "../../../../node_modules/bootstrap/scss/navbar";
-//@import "../../../../node_modules/bootstrap/scss/card";
+@import "../../../../node_modules/bootstrap/scss/card";
 //@import "../../../../node_modules/bootstrap/scss/breadcrumb";
 //@import "../../../../node_modules/bootstrap/scss/pagination";
 @import "../../../../node_modules/bootstrap/scss/badge";
@@ -29,12 +30,22 @@
 //@import "../../../../node_modules/bootstrap/scss/media";
 @import "../../../../node_modules/bootstrap/scss/list-group";
 @import "../../../../node_modules/bootstrap/scss/close";
-//@import "../../../../node_modules/bootstrap/scss/modal";
+@import "../../../../node_modules/bootstrap/scss/modal";
 //@import "../../../node_modules/bootstrap/scss/tooltip";
 @import "../../../../node_modules/bootstrap/scss/popover";
 //@import "../../../../node_modules/bootstrap/scss/carousel";
 @import "../../../../node_modules/bootstrap/scss/utilities";
 
-@import "fonts.scss";
+//@import "../../../../node_modules/jquery.mmenu/dist/jquery.mmenu.all";
+@import "../../../../node_modules/corenav/coreNavigation";
+
+@import "type";
+@import "grid";
+@import "buttons";
+@import "fonts";
 @import "common";
 @import "layout";
+@import "images";
+@import "tags";
+@import "homepage";
+@import "sections";
