+++ src/main/java/com/pinwheel/anabel/Anabel.java
@@ -2,7 +2,6 @@ package com.pinwheel.anabel;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 /**
