+++ src/main/java/com/pinwheel/anabel/controller/HomeController.java
@@ -21,7 +21,7 @@ public class HomeController {
      */
     @GetMapping("/")
     public String home() {
-        return "home";
+        return "homepage/home";
     }
 
     /**
