+++ src/main/resources/i18n/general.properties
@@ -1,0 +1,9 @@
+general.site.name   =Anabel
+general.languages   =Languages
+general.languages.en=English
+general.languages.ru=Русский
+
+# Articles, particles and pretexts
+particle.and        =and
+particle.or         =or
+particle.the        =the
