+++ src/main/resources/i18n/general_ru.properties
@@ -1,0 +1,9 @@
+general.site.name   =Анабель
+general.languages   =Языки
+general.languages.en=English
+general.languages.ru=Русский
+
+# Articles, particles and pretexts
+particle.and        =и
+particle.or         =или
+particle.the        =
