+++ src/main/resources/i18n/messages.properties
@@ -1,47 +1,92 @@
-general.site.name                           =Anabel
-general.languages                           =Languages
-general.languages.en                        =English
-general.languages.ru                        =Русский
+footer.link.header.contacts                   =Contacts
+footer.link.header.to-read                    =To read
 
-footer.link.header.contacts                 =Contacts
-footer.link.header.to-read                  =To read
+footer.link.write.us.0                        =If you can program or have other useful skills and want to join our team
+footer.link.write.us.1                        =write to us
+footer.link.write.us.2                        =Perhaps you are the one we need
+footer.link.report-bug                        =Report a bug
+footer.link.advertise                         =Advertise
+footer.link.violation.rights                  =Rights violation
+footer.link.about.project                     =About anabel
+footer.link.about.company                     =About company
+footer.link.terms-of-user                     =Terms of use
+footer.link.privacy-policy                    =Privacy
 
-footer.link.write.us.0                      =If you can program or have other useful skills and want to join our team
-footer.link.write.us.1                      =write to us
-footer.link.write.us.2                      =Perhaps you are the one we need
-footer.link.report-bug                      =Report a bug
-footer.link.advertise                       =Advertise
-footer.link.violation.rights                =Rights violation
-footer.link.about.project                   =About anabel
-footer.link.about.company                   =About company
-footer.link.terms-of-user                   =Terms of use
-footer.link.privacy-policy                  =Privacy
+# Home page
+homepage.banner.first.general.text            =Anabel is a freelance services exchange that provides quality \
+  intermediation between customers and performers.
+homepage.banner.first.client.heading          =For customers
+homepage.banner.first.client.text             =We will select the best performers to meet your requirements. Rating \
+  system and reviews will help minimize risks. We also provide a special selection
+homepage.banner.first.client.text2            =system for students
+homepage.banner.first.client.button           =Order a job
+homepage.banner.first.performer.heading       =For performers
+homepage.banner.first.performer.text          =Perform orders remotely without commissions and service charges. Large \
+  selection of destinations and convenient tenders.
+homepage.banner.first.performer.button        =Become a performer
 
-link.account-already-exist                  =Already has account?
-link.account-not-exist-yet                  =Doesn't have account?
-link.account-profile                        =Profile
+homepage.section.first.heading                =Popular directions
+homepage.section.first.technology.heading     =Technology
+homepage.section.first.technology.text        =Each task can be tagged for easy searching. Tags describe the areas of \
+  this work: programming languages, technologies, programs used, etc. You can create your own tag if \
+  it did not exist before and track its activity.
+homepage.section.first.establishments.text    =Each task can be tagged for easy searching. Tags describe the areas of \
+  this work: programming languages, technologies, programs used, etc. You can create your own tag if \
+  it did not exist before and track its activity.
+homepage.section.first.establishments.heading =Establishments
+homepage.section.first.link.technology        =All Technology
+homepage.section.first.link.establishments    =All Establishments
 
-form.user.heading.create-account            =Create your Anabel account. It’s free and only takes a minute.
-form.user.heading.sign-in                   =Sign in for additional features.
+homepage.banner.second.commitment.heading     =Commitment system
+homepage.banner.second.commitment.text        =When writing a technical task, the customer can indicate the \
+  commitments that must be fulfilled by the
+homepage.banner.second.commitment.text2       =customer
+homepage.banner.second.commitment.text3       =performer
+homepage.banner.second.commitment.text4       =deadline, amount of work, number of attracted users, etc.
+homepage.banner.second.commitment.ps.text     =You can look at the complete list of the commitments on our
+homepage.banner.second.commitment.ps.text2    =learning page
 
-form.user.display-name.label                =Display name
-form.user.email.label                       =Email
-form.user.first-name.label                  =First name
-form.user.last-name.label                   =Last name
-form.user.password.label                    =Password
-form.user.confirm-password.label            =Confirm password
-form.user.sign-in.label                     =Sign In
-form.user.sign-up.label                     =Sign Up
-form.user.logout.label                      =Sign Out
+homepage.section.second.heading               =Find the right performer
+homepage.section.second.text                  =List of the best performers for the current year
+homepage.section.second.performer.button      =Connect
+homepage.section.second.link.performers-search=The entire list of artists and search for them
+
+# Links
+
+link.account-already-exist                    =Already has account?
+link.account-not-exist-yet                    =Doesn't have account?
+link.account-profile                          =Profile
+
+# Form headings
+form.user.heading.create-account              =Create your Anabel account. It’s free and only takes a minute.
+form.user.heading.sign-in                     =Sign in for additional features.
+
+form.section.heading.create                   =Section Creation
+
+# Form labels
+form.user.display-name.label                  =Display name
+form.user.email.label                         =Email
+form.user.first-name.label                    =First name
+form.user.last-name.label                     =Last name
+form.user.password.label                      =Password
+form.user.confirm-password.label              =Confirm password
+form.user.sign-in.label                       =Sign In
+form.user.sign-up.label                       =Sign Up
+form.user.logout.label                        =Sign Out
+
+form.section.name.label                       =Section name
+form.section.add.label                        =Add section
+form.section.cancel.label                     =Cancel
+
+auth.message.bad-credentials                  =Bad credentials.
+auth.message.disabled                         =This user is not available.
+auth.message.expired                          =This user credentials are expired.
+auth.message.blocked                          =This user are blocked.
+auth.message.captcha.empty                    =Fill captcha.
+
+auth.register.activation.verify               =Verify your email for completing registration.
+auth.register.activation.expired              =Activation code is expired. {0}.
+auth.register.activation.expired.link-phrase  =Resend mail
 
 
-auth.message.bad-credentials                =Bad credentials.
-auth.message.disabled                       =This user is not available.
-auth.message.expired                        =This user credentials are expired.
-auth.message.blocked                        =This user are blocked.
-auth.message.captcha.empty                  =Fill captcha.
-
-auth.register.activation.verify             =Verify your email for completing registration.
-auth.register.activation.expired            =Activation code is expired. {0}.
-auth.register.activation.expired.link-phrase=Resend mail
