+++ src/main/resources/i18n/messages_ru.properties
@@ -1,13 +1,8 @@
-general.site.name                           =Анабель
-general.languages                           =Языки
-general.languages.en                        =English
-general.languages.ru                        =Русский
-
 footer.link.header.contacts                 =Обратная связь
 footer.link.header.to-read                  =Информация
 
-footer.link.write.us.0                      =Если вы умеете программировать или имеете другие полезные навыки и хотите присоединиться к \
-  нашей команде
+footer.link.write.us.0                      =Если вы умеете программировать или имеете другие полезные навыки и хотите \
+  присоединиться к нашей команде
 footer.link.write.us.1                      =напишите нам
 footer.link.write.us.2                      =Возможно вы тот, кто нам нужен
 footer.link.report-bug                      =Сообщить об ошибке
@@ -17,6 +12,47 @@ footer.link.about.project                   =Про Анабель
 footer.link.about.company                   =Про компанию
 footer.link.terms-of-user                   =Условия использования
 footer.link.privacy-policy                  =Политика конфиденциальности
+
+# Home page
+homepage.banner.first.general.text          =Анабель - это биржа фриланс услуг, которая обеспечивает качественное \
+  посредничество между заказчиками и исполнителями.
+homepage.banner.first.client.heading        =Для клиентов
+homepage.banner.first.client.text           =Мы подберём лучших исполнителей для удовлетворения ваших требований. Рейтинговая \
+  система и отзывы помогут минимизировать риски. Также мы предоставляем специальную 
+homepage.banner.first.client.text2=систему подбора для студентов
+homepage.banner.first.client.button         =Заказать работу
+homepage.banner.first.performer.heading     =Для исполнителей
+homepage.banner.first.performer.text        =Выполняйте заказы удалённо без комиссий и сервисных сборов. Большой выбор \
+  направлений и удобные тендеры.
+homepage.banner.first.performer.button      =Стать исполнителем
+
+homepage.section.first.heading=Популярные направления
+homepage.section.first.technology.heading=Технологии
+homepage.section.first.technology.text=Каждая задача может быть помечена тегами, для удобного  поиска. Теги описывают \
+  направления данной работы: языки программирования, технолонии, используемые программы и т.д.. \
+  Вы можете создать свой тег, если он раньше не существовал и отслеживать его активность.
+homepage.section.first.establishments.text=Каждая задача может быть помечена тегами, для удобного  поиска. Теги описывают \
+  направления данной работы: языки программирования, технолонии, используемые программы и т.д.. \
+  Вы можете создать свой тег, если он раньше не существовал и отслеживать его активность.
+homepage.section.first.establishments.heading=Образование
+homepage.section.first.link.technology=Все технологии
+homepage.section.first.link.establishments=Все учебные заведения
+
+homepage.banner.second.commitment.heading=Система обязательств
+homepage.banner.second.commitment.text=При написании технического задания заказчик может указать \
+  обязательства, которые должны быть выполнены со стороны
+homepage.banner.second.commitment.text2=заказчика
+homepage.banner.second.commitment.text3=исполнителя
+homepage.banner.second.commitment.text4=срок выполнения, объём работы, количество привлечённых клиентов и т.д.
+homepage.banner.second.commitment.ps.text=Полный список обязательств вы можете посмотреть на нашей
+homepage.banner.second.commitment.ps.text2=обучающей странице
+
+homepage.section.second.heading=Найди правильного специалиста
+homepage.section.second.text=Список лучших исполнителей за текущий год
+homepage.section.second.performer.button=Связаться
+homepage.section.second.link.performers-search=Весь список исполнителей и поиск по ним
+
+# Links
 
 link.account-already-exist                  =Уже есть аккаунт?
 link.account-not-exist-yet                  =Ещё нет аккаунта?
