--- src/main/resources/templates/home.ftl
@@ -1,5 +1,0 @@
-<#import "layout/main.ftl" as m>
-
-<@m.page>
-    Home page
-</@m.page>
