+++ src/main/resources/templates/layout/nolanding.ftl
@@ -1,23 +1,21 @@
 <#macro page>
-<html>
+    <html>
     <#include "head.ftl">
     <body>
     <div id="app">
         <#include "header.ftl">
         <main class="flex-grow-1">
             <#include "alert.ftl">
-
-            <div class="container-fluid mt-3">
+            <div class="container mt-3">
                 <div class="row">
-                    <div class="offset-2 col-3">
-                        <#nested>
-                    </div>
+                    <#nested>
                 </div>
             </div>
+
         </main>
         <#include "footer.ftl">
         <#include "scripts.ftl">
     </div>
     </body>
-</html>
+    </html>
 </#macro>
