+++ src/main/resources/templates/static/cookie_en.ftl
@@ -1,4 +1,4 @@
-<#import "../layout/main.ftl" as m>
+<#import "../layout/landing.ftl" as m>
 <#import "/spring.ftl" as spring />
 
 <@m.page>
